refactor(parser): split parser.rs into parser/ module
Break the 3038-line parser.rs into eight focused files grouped by
grammar production. Largest slice is expressions.rs at 963 lines —
all under the ~1200-line target from issue #41.

  error.rs       (37)   ParserError + Display/Error impls
  state.rs      (318)   Parser struct, new(), parse(), token helpers
  expressions.rs (963)  precedence climbing, primary, list/dict/closure,
                        arg/param lists
  statements.rs (709)   let/var, if/for/while/match, parallel, try/catch,
                        select, guard/require/deadline/yield/mutex/defer
  decls.rs      (546)   pipeline, import, attributes, fn/tool/type/enum/
                        struct/interface/impl
  types.rs      (225)   type expr, type params, where clauses, shape types
  patterns.rs    (94)   binding pattern parsing
  mod.rs        (207)   module decls, pub use, tests

Public API (Parser, ParserError, parse, parse_single_expression,
all_errors, new) is unchanged. All conformance fixtures parse
identically (472/472 pass); parser unit tests (164/164) and
tree-sitter tests (16/16) also green.

Closes #41
